AnimalPredVercel/requirements.txt
dvc==2.53.0
Flask==2.2.3
joblib==1.2.0
keras==2.12.0
numpy==1.23.4
pandas==1.5.1
Pillow==9.5.0
Werkzeug==2.2.3
tensorflow==2.12.0
